Refactored components
Removed Days component from Home.
Makes more sense to just have the temperature, day as a part of a home component
