+++ weather-app/src/components/Days.js
@@ -1,17 +1,8 @@
 import React from 'react';
-import Day from './Day';
-import {getDate} from '../utils/date.js'
-import '../styling/home.css';
 
 const Days = (props) => {
-
-        // get date
-        const date = getDate();
-
         return(
         <div className="days">
-            <h1>{date}</h1>
-            <Day temp={props.temp} desc={props.desc}/>
         </div>
     );
 };
