+++ weather-app/src/media/forecast-bg.jpg
