+++ weather-app/src/styling/App.css
@@ -1,5 +1,5 @@
 body{
-    background-image: url("../media/home-background.jpg");
+    background-image: url("../media/forecast-bg.jpg");
     background-repeat: no-repeat;
     text-align: center;
     font-family: 'Open Sans', sans-serif;
